# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample fiber project server.",
        "title": "Machine task",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/admin/job": {
            "get": {
                "security": [
                    {
                        "authorization": []
                    }
                ],
                "description": "Retrieve a list of jobs with pagination and optional search",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job Management Admin"
                ],
                "summary": "Get jobs",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 5,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved jobs",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "authorization": []
                    }
                ],
                "description": "Create a new job posting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job Management Admin"
                ],
                "summary": "Create a job",
                "parameters": [
                    {
                        "description": "Job Create",
                        "name": "job",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodels.CreateJob"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created job",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "authorization": []
                    }
                ],
                "description": "Delete an existing job posting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job Management Admin"
                ],
                "summary": "Delete a job",
                "parameters": [
                    {
                        "description": "Job Delete",
                        "name": "job",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodels.DeleteJob"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted job",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    }
                }
            }
        },
        "/admin/job/details": {
            "get": {
                "security": [
                    {
                        "authorization": []
                    }
                ],
                "description": "Retrieve details of a specific job with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job Management Admin"
                ],
                "summary": "Get complete details of job",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 5,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Job ID",
                        "name": "jobid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved job details",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    }
                }
            }
        },
...
        "/applied": {
            "get": {
                "security": [
                    {
                        "authorization": []
                    }
                ],
                "description": "Retrieve a list of applied jobs with pagination and optional search",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apply job user"
                ],
                "summary": "Get applied jobs",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 5,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved applied jobs",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "authorization": []
                    }
                ],
                "description": "Submit a job application",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apply job user"
                ],
                "summary": "Apply for a job",
                "parameters": [
                    {
                        "description": "Job Application",
                        "name": "jobApplication",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestmodels.JobApplication"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully applied for the job",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    }
                }
            }
        },
        "/job": {
            "get": {
                "security": [
                    {
                        "authorization": []
                    }
                ],
                "description": "Retrieve a list of jobs with pagination and optional search",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user job"
                ],
                "summary": "Get jobs",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 5,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved jobs",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    }
                }
            }
        },
...
        "/resume": {
            "post": {
                "security": [
                    {
                        "authorization": []
                    }
                ],
                "description": "Upload a resume file and decode its content",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user job"
                ],
                "summary": "Upload and decode resume",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Resume file",
                        "name": "resume",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resume successfully added",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/responsemodels.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requestmodels.CreateJob": {
            "type": "object",
            "required": [
                "company_name",
                "description",
                "title"
            ],
            "properties": {
                "company_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "requestmodels.DeleteJob": {
            "type": "object",
            "properties": {
                "jobID": {
                    "type": "string"
                }
            }
        },
        "requestmodels.JobApplication": {
            "type": "object",
            "required": [
                "job_application_id"
            ],
            "properties": {
                "job_application_id": {
                    "type": "string"
                }
            }
        },
...
        "responsemodels.Response": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                },
                "result": {},
                "status_code": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "authorization": {
            "description": "IMPORTANT: TYPE \"Bearer\" FOLLOWED BY A SPACE AND JWT TOKEN.",
            "type": "apiKey",
            "name": "authorization",
            "in": "header"
        }
    }
}